drm/i915/gem: Fix nesting of parallel context locks for arbitrary depth
The nesting depth with a parallel context is ultimately limited by the
number of engines the user combines into a single parallel context. This
can easily exceed lockdep's MAX_LOCKDEP_SUBCLASSES triggering warnings
and disabling lockdep.

The approach to handle arbitrary nesting to use mutex_lock_nest_lock,
and strictly use an outer lock (the parallel parent context) as a
guard for nesting all the inner locks (the parallel children).

Fixes: 544460c33821 ("drm/i915: Multi-BB execbuf")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>