drm/i915: Force a GuC reset after driver test failure
Between tests, IGT flushes and serialises with outstanding work. In
doing so, if the system is unresponsive, we forcibly reset the device.

Currently, we are only checking if there is outstanding activity on the
GT engines, but ignore if there is an ongoing GuC communication issue.
Include flushing the GuC communications with the wait and decision to
reset if that flush times out.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>