drm/i915: Add idleness check in i915 GuC selftest.
The perf selftest is keeping the context pinned while
checking for idle. This does not work with GuC submission,
because in this mode the idleness is determined based on
the wakeref count and a pinned context owns a wakeref.
Instead of relying only on the check for idle, we can
handle this case by checking that there are no requests in
flight.

Signed-off-by: Zhanjun Dong <zhanjun.dong@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>