INTEL_DII: drm/i915/gt: Add a selftest to investigate PTE write-tearing
The objective is to determine whether or not the HW is susceptive to the
effect of write tearing of the PTE updates. That is whether is may
sample the PTE in the middle of the host updating the entry and so
attempt to load an invalid physical address. To achieve this we launch a
spining batch buffer that continually invalidates its TLB cache and
attempts to load a value from memory and compare that value against
expected. If we ever see an unexpected value or generate a CAT error, we
know the HW has gone astray.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>