INTEL_DII: drm/i915: Allow dynamic modification of masked regs
For the upcoming dynamic sysfs toggle patch, some pre-existing
masked register need to be dynamically set and clear. For this to work on
clearing, we need to be careful to keep the upper mask but clear the
relevant bits on the lower.

Testcase: igt/i915_selftest --r live --dynamic=debugger

Cc: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Cc: Christoph Manszewski <christoph.manszewski@intel.com>
Cc: Chris Wilson <chris.p.wilson@intel.com>
Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>