INTEL_DII/FIXME: drm/i915/gsc: MTL: Enable driver-FLR for GSC workaround
On MTL, since GSC firmware is using dynamically allocated memory,
enable the driver-FLR triggers (that happens at gt-init / fini)
to reset and quiesce the GSC microntroller and firmware at driver
unload.

This is the only way to ensure the GSC firmware doesn't continue
operation after driver unload. This is important on MTL because
when i915 loads the GSC firmware, it provides memory that was
allocated from the kernel and that memory would eventually be
returned without a way to cease unrelated GSC engine operation
that can still access that memory.

For now, enable this by default on MTL via a device info
config. Continue to allow the "force_driver_flr" modparam
to override that param off even for MTL because depending on
the current pre-PV MTL IFWIs we receive do occasionally break
and in such cases we want to skip the driver-FLR. For example,
in the past we saw entire MMIO register stop working (always
returning 0xFFFFFFFF with a broken IFWI).

NOTE: DriverFLR is not the same as PCI config space FLR,
but it is the name chosen by Bspec: 53111, 53113.

FIXME: Drop this patch once MTL GSC fw uses stolen memory

v2: Move param sanitize to seperate function in i915_driver.c

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>