drm/i915/vf: Disable tasklets during post-migration fixups
While the worker is performing fixups after migration, tasklets
must be deactivated. Otherwise, that would cause unsubmitted
requests to be submitted again on a still non-functional hardware.

Unsubmitted requests are scheduled in queue to be re-submitted when
the post-migration worker is active, so tasklets need to be
disabled to avoid the premature submissions. That differentiates
the migration process from suspend/resume process, in which
submission queue is flushed and cleared of any remains.
The tasklet there wasn't a problem, as it was assured there
is nothing to submit.

Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Signed-off-by: Tomasz Lis <tomasz.lis@intel.com>