INTEL_DII: drm/i915: Update GEM_CREATE to support BO segments
This is first part of changes to implement a reduced in scope version
of finer-grained 'chunk' BO granularity.  Here, we implement support to
divide the BO into 'fixed size' segments.  They are stored as a list in
the BO, and are of the same type (drm_i915_gem_object).
BO segments are never 'published' and so are not explicitly freed by
user. We just need to decrement the reference count for segments when
the parent BO is being freed.

By using drm_i915_gem_object data structure to describe the individual
backing store segments, we will see in follow-on patches, how easy it
becomes to support this smaller size granularity for device page faults
and migration at 'chunk' granularity.

Currently allow minimum chunk size of 2M as smaller is not really worth
the overhead of using many i915_gem_objects and not the intent of this
feature.  However, may relax this in future.

Here we automatically create segments if size meets our criteria
(size > 2 chunks).  But in the very next patch in this series we add
GEM_CREATE extension for user to request chunking of BOs.

Add a few GEM_BUG_ON in shrinker/eviction/migration paths to confirm that
these only find the segment BOs (which have the backing store).

v2: convert from array to list, increased allowed number of chunks
v3: rename function that computes segment_size (Oak)
    use u64 instead of size_t (Chris)
v4: remove MAX_CHUNKS and dynamic roundup of chunk_size. If we do this
    then we need to extend uAPI to return chunk size to user. Per Chris,
    let's not constrain or alter the users' request (Chris/Jaime)
    move release of segments into helper and use it to fix teardown
    on error path (Chris), other tidyness (Chris/Mike)
    guard that we don't attempt to migrate segment BO's parent (Chris)
v5: per UMD request, update gem_create_trace to print # of segments (Brian)

Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Reviewed-by: Oak Zeng <oak.zeng@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>