INTEL_DII/FIXME: Dump error capture on default context timeout
When waiting for gt idle and using GuC, we are waiting for all the
messages to go through. If there is a failure here, the possible debug
is to look at the GuC log to understand what it was doing. However there
is no way to get the log debug later if it's not dumped at this point.

FIXME: "Dump error capture to kernel log" is marked as NOT_UPSTREAM. So
if if this is useful upstream that support needs to be cleaned up and
posted before this.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>