drm/i915/gem: Track object migration
Instantiating the backing store for an object may not be instantaneous.
For example, upon construction we need to ensure that the contents of
the backing store is clear, and at other times we may need to repopulate
the backing store with the contents from a swap buffer. To accommodate
handling this in an asynchronous manner, we want to track the fence for
the operation and also track any persistent error that may arise.
Similarly, before using the backing store, we must serialise with the
migration fence and check the error status.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>