INTEL_DII/DO_NOT_UPSTREAM: drm/i915/pxp: Pseudo-backport interfaces for protected objects
This patch is an attempt to recreate the upstream patch:

commit d3ac8d42168a ("drm/i915/pxp: interfaces for using protected
objects")

A cherry-pick or backport has been eluding us because it's
a much larger ongoing effort: On the upstream repo, PXP feature
got merged in after proto-context and gem-create-ext redesigns
(and after a decent amount of refactoring for mem-region support).
Thus backporting all the dependencies is a mega task on its own
and is an ongoing work-in-progress via VLK-29839

In the meantime, in order to prepare the path to remove PRELIM
UAPIs for PXP (which is not being used by any Intel customer)
and prepare DII tree for MTL Single-Session-PXP design changes
(that will be driven from upstream first), we need to start
making the rest of the PXP on DII look like and behave like
its upstream counterpart.

A snippet of the upstream commit message for the patch that is
being replicated here is:

   This api allow user mode to create protected buffers and to mark
   contexts as making use of such objects. Only when using contexts
   marked in such a way is the execution guaranteed to work as expected.

   Contexts can only be marked as using protected content at creation time
   (i.e. the parameter is immutable) and they must be both bannable and
   not recoverable. Given that the protected session gets invalidated on
   suspend, contexts created this way hold a runtime pm wakeref until
   they're either destroyed or invalidated.

   All protected objects and contexts will be considered invalid when the
   PXP session is destroyed and all new submissions using them will be
   rejected. All intel contexts within the invalidated gem contexts will
   be marked banned. Userspace can detect that an invalidation has occurred
   via the RESET_STATS ioctl, where we report it the same way as a ban due
   to a hang.

FIXME: This patch needs to be removed when VLK-29839 is completed
and we can backport the original patch from upstream
(d3ac8d42168a9be7380be8035df8b6d3780ec2a1).

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>