INTEL_DII/NOT_UPSTREAM/UAPI: drm/i915/ats-m: Exporting uid for ATS-M via sysfs
This patch exports uid to the userspace via sysfs.

(v2)
-Read uid values during driver initialization.(Aravind)
-Drop usage of show_uid to determine ATSM from device_info.(Aravind)
-Add atsm check.(Aravind)
-use smaller drm_ over DRM.(Aravind)
-Use sysfs_file_create for single file.(Ashutosh)

(v3)
-Remove drm_info print for uid.(Aravind)
-Move registers definition to i915_reg.h.(Aravind)
-Move uid to struct intel_runtime_info.(Aravind)
-Drop atsm check based on pciid.(Aravind/chris)
-Add ATSM as a subplatform of DG2.(Chris)

(v4)
-Drop ATSM as subplatform of DG2.(Matt)
-Use device_info to recognize ATS_M.(Matt)

(v5)
-Change register names to CSC_DEVUID_LWORD/HWORD.(Joonas)
-Use feature name has_csc_uid in device_info. (Joonas)

(v6)
-Add prelim_ prefix to sysfs entry.(Ashutosh)

FIXME: Patch needs to be dropped, if there is an agreement to do CSC change
to allow reading the UID directly via MEI on ATS-M.
If there is no such agreement, we should ask if there is agreement to
support the UID register on wide set of platforms to allow upstreaming.

Signed-off-by: Himal Prasad Ghimiray <himal.prasad.ghimiray@intel.com>
Reviewed-by: Aravind Iddamsetty <aravind.iddamsetty@intel.com>