INTEL_DII: drm/i915: Error for ioctls that don't support segmented BOs
Following ioctls will return errors if used with segmented BOs:
GEM_SET_CACHING
GEM_SET_DOMAIN
GEM_EXECUFFER2 (for exec objects)

SET_CACHING and SET_DOMAIN are disabled for DGFX in upstream, but are
still supported in DII.
Level0 confirmed that these are only being used for integrated GFX.
As DII is essentially ATS-M and PVC only, these are not priority to
support.  Can consider to enable as follow-on.

v5: check for unsupported should be upfront (Mike)

Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>