INTEL_DII: drm/i915: Update vm_unbind_obj to support segmented BOs
For segmented BOs: during VM_BIND, we divided the VA range and created many
smaller VMAs.  But this breaks VM_UNBIND.

When user requests to UNBIND, we now need to perform an adjacency search
in order to find and unbind the set of VMAs for that VA range. We use the
adjacency list which was constructed during VM_BIND.

Partial unbind is not a supported feature, so care is taken here to not
allow unbinding of subset of the VMAs that correspond to original address
range.

Unbind is done using multiple workers (one per VMA). It might be simpler to
use a single worker to do the unbind for all the VMAs in the adjacency list.
But my understanding is that we need a dma_fence_worker per VMA in order to
not break and properly advance the vma->active timeline.

v2: Rebase on top of unbind worker changes from Chris.
v3: refactor as adjacency list is built upfront now during VM_BIND
v4: move logic for unbind verification of segments into helper (Oak)
v5: remove unneeded use of vma_next (Oak)

Reviewed-by: Oak Zeng <oak.zeng@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>