INTEL_DII/UAPI/NOT_UPSTREAM: drm/i915/pxp: Add object PRELIM UAPI
This patch adds the PRELIM definitions for PXP buffer
creation using DII's GEM_CREATE overload of EXT param.

NOTE: Part of PXP-Pseudo-backport, we moved as much PXP
UAPI behaviors as we could to pending-upstream section.
However, that move required dropping PRELIM definitions.
But as we move up the DII pile, we use the GEM_CREATE
overload of EXT param as the default method for user space.

Thus add just the PRELIM definitions where required to
ensure user space can continue using it with the same
consistent behavior as the upstream PXP buffer creation.

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>