INTEL_DII: gt/intel_engine_cs: Check if GSC enabled before init
When the system is run with HuC and GSC disabled (due to FLR having issues
on the CI machines), the GSC engine cannot reset.  This is because the GSC
engine requires the GSC FW in order to reset, and attempting to reset the
GSC without this requirement can taint the kernel.

A suggested solution to this quandry is to skip GSC engine initialization
when the GSC FW is disabled.  This will prevent said engine from being
included in any tests that may require an engine reset when engine resets
for GSC are not possible.

v2:
- The GSC engine must also be removed from the engine mask in this case.

v3:
- Since the mask is updated to drop the GSC engine when the conditions
are correct, the additional condition check in the init is no longer
required, as the engine will be skipped by the mask.

References: VLK-38976

Signed-off-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Suggested-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
CC: Matt Roper <matthew.d.roper@intel.com>
CC: Chris Wilson <chris.p.wilson@linux.intel.com>
CC: John C Harrison <John.C.Harrison@intel.com>
Reviewed-by: Stuart Summers <stuart.summers@intel.com>
Acked-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>