INTEL_DII: drm/i915: Propagate VM_ADVISE hints to BO segments
We need the VM_ADVISE hints to be set in the segments so that
the migration policy implemented in device page fault handler
is applied to BO segments.
On error in setting hints for any of the segments, we revert
hint changes and rollback to the prior hint values so that
applied hints are consistent.

v4: this series reverted to using individual lock per segment,
    thus we lock each segment in i915_gem_object_set_hint
v5: on error, hints are rolled back to prior value (Mike)

Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>