INTEL-DII: drm/i915/pvc: Bump pre-emption timeout for BCS engines
If all copy engines are thrashing away at the PCIe bus, plus all
compute engines doing the same, then the per copy bandwidth becomes
tiny. That results in copy operations that should take milliseconds
actually taking seconds. So bump the pre-emption timeout to allow for
the worst case scenario (currently believed to be around 2s).

Note that the responsiblity is on the UMD to not submit large copy
operations. The limit given here is based on all copies being limited
to 8MB size. Anything larger will still run the risk of being killed
due to pre-emption timeout on a heavily loaded system.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>