INTEL_DII/FIXME: drm/i915/: Fix NPD in PMU during driver teardown
In the driver teardown, we are unregistering the gt prior
to unregistering the PMU. This means there is a small window
of time in which the application can request metrics from the
PMU, some of which are calling into the uapi engines list,
while the engines are not available. In this case we can
see null pointer dereferences.

Fix this ordering in both the driver load and unload sequences.

Additionally add a check for engine presence to prevent this
NPD in the event this ordering is accidentally reversed. Print
a debug message indicating when they aren't available.

(cherry picked from commit 1491b2707b092f324e05d694886d4aa82cec50c9)

v1: Actually address the driver load/unload ordering issue
v2: Use drm_WARN_ONCE instead of a debug print

Signed-off-by: Stuart Summers <stuart.summers@intel.com>
Acked-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Acked-by: Anshuman Gupta <anshuman.gupta@intel.com>