INTEL_DII: drm/color: Add pipe degamma mode property
Add degamma mode property for CRTC. This is an enum property with
values as blob_id's. The blob that the blob id refers to is used
to pass information regarding the mode supported and the range of
degamma mode supported with number of lut coefficients to the user
space. The userspace can select any of the supported degamma mode
using this enum property.

v2: Unify gamma and degamma attach property (Bhanu)

Signed-off-by: Borah, Chaitanya Kumar <chaitanya.kumar.borah@intel.com>
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>