UPSTREAM: ASoC: Intel: avs: Introduce debug-context aware helpers
Debug-related fields and log-dumping are useful when debugfs is enabled.
Define them under CONFIG_DEBUG_FS and provide stubs when the config is
disabled so that the code that makes use of these needs not to be
complicated unnecessarily.

Members that are duplicated by this patch will be removed by the follow
up changes.

Signed-off-by: Cezary Rojewski <cezary.rojewski@intel.com>
Link: https://lore.kernel.org/r/20221202152841.672536-7-cezary.rojewski@intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
Change-Id: I90e67aadeaa0162cb80f771b3f6d7c1c441e334c
(cherry picked from commit 9e3c15beb8976771f95ba30b3da8bd35dc7188ac)
Signed-off-by: Cezary Rojewski <cezary.rojewski@intel.com>